Added a mechanism to apply preferences
This is meant to be used as a way to transfer preferences from one
computer to the next. Each application must deal with its preferences in
its own way, which means that there is no "right" way to handle
preferences at the moment. Just save some file(s) on the "preferences/"
directory and use them on the "setup-preferences.sh" script to import
them into the current machine.